requirements.txt
selenium
pytest
pytest-bdd
allure-pytest
requests